requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-appconf==1.0.6
django-classy-tags==4.1.0
django-cookie-consent==0.5.0b0
django-taggit==5.0.1
docutils==0.20.1
idna==3.6
imagesize==1.4.1
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
psycopg==3.1.16
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.9.0
update==0.0.1
urllib3==2.1.0
